By comparing output from the example Fortran implementation of Lucifer against my own implementation, I was able to uncover some bugs. After this commit, my implementation seems to produce the same output as the Fortran implementation.
1. There are some quirks regarding how s-boxes are applied, that are not clear in the algorithm's description, but are explicit in the example Fortran implementation given. First, the 4-bit nibbles ("hexdigits") of the message byte are interpreted with the RIGHT-MOST bit being the most significant, before looking up the corresponding value in the appropriate S-box. Second, the entire byte is reversed after the outputs of the S-boxes are put together again.

2. The Fortran implementation and my implementation both swap pointers to different halves of the message between rounds rather than exchanging the contents of each half. However, the Fortran implementation also swaps the contents once at the end. It is not clear to me why this is necessary, but I have added it in order to be compatible. The explanations given in the paper are:

"Because the subroutine operates on the message in place, it is necessary to physically swap the contents of the upper and lower halves of the after the end of the sixteen rounds in order to have the halves in the correct order."

and this comment:

"physically swap upper and lower halves of the message after the last round. we wouldn't have needed to do this if we had been swapping all along."

My testing seems to indicate that physically swapping halves without the swap at the end DOES NOT produce the same results as pointer-swapping with one physical swap at the end. In fact, physical swapping and pointer swapping (both without any swap at the end) produce the same results. This makes sense, because any even number of swaps should end up with the (modified) halves in their original position. In the case of Lucifer, there are 16 swaps, one for each round.